third_party/blink/web_tests/external/wpt/html/browsers/history/the-history-interface/001-expected.txt

This is a testharness.js-based test.
[FAIL] pushState must remove any tasks queued by the history traversal task source
  assert_equals: expected "test4" but got "test"
Harness: the test ran to completion.
